Remove success flash when updating brief response
We believe that this success flash message is misleading some suppliers. We have some evidence to suggest that they think the flash message means that they've successfully submitted their brief response. This means that they don't press the button to submit their application.

Our evidence here is not definitive - we've got a pattern of support requests, but we've not done any user research or proper analysis.

We don't need this flash, however. Users can see on the page that their information has been updated.

For this reason, we're just going to go ahead and remove the flash. We'll check back in a few weeks to see what effect it's had on the rate of submitted brief responses and support requests.